requirements.txt
torch
transformers==4.12.5
datasets
sacrebleu
Pillow
flask
onnxruntime
vncorenlp
flask_sqlalchemy
librosa
matplotlib
nltk
tensorboard
tensorboardX
scipy
sklearn
torchvision
pyctcdecode